pyhalflife/requirements.txt
pycapnp==1.1.0
